+++ frontend/app/shoulder.tsx
@@ -1,0 +1,9 @@
+interface Props {
+  children?: React.ReactNode
+}
+
+export default function Shoulder({ children }: Props) {
+  return (
+    <section className="flex-col w-64 gap-2 hidden sm:flex">{children}</section>
+  )
+}
