--- frontend/components/elements/index.ts
@@ -1,5 +1,0 @@
-import Button from "./Button/Button"
-import Link from "./Link/Link"
-import Input from "./Input/Input"
-
-export { Button, Link, Input }
