+++ frontend/components/feed/index.ts
@@ -1,0 +1,2 @@
+import Feed from "./feed"
+export default Feed
