+++ frontend/components/link/index.ts
@@ -1,0 +1,2 @@
+import Link from "./link"
+export default Link
