+++ frontend/components/post-preview/post-title.tsx
@@ -1,0 +1,9 @@
+import Link from "@/components/link"
+
+export default function Title({ title, url }: { title: string; url: string }) {
+	return (
+		<Link href={url} className="font-medium text-zinc-50">
+			{title}
+		</Link>
+	)
+}
