--- frontend/components/segments/Shoulder/Shoulder.tsx
@@ -1,11 +1,0 @@
-interface Props {
-	children?: React.ReactNode
-}
-
-export default function Shoulder({ children }: Props) {
-	return (
-		<section className="flex flex-col w-64 gap-2 hidden sm:flex">
-			{children}
-		</section>
-	)
-}
