--- frontend/components/segments/Shoulder/index.ts
@@ -1,2 +1,0 @@
-import Shoulder from "./Shoulder"
-export default Shoulder
