+++ frontend/components/tabs/index.ts
@@ -1,0 +1,2 @@
+import Tabs from "./tabs"
+export default Tabs
