--- frontend/lib/feed/Content.ts
@@ -1,6 +1,0 @@
-import { Feed, Post } from "@/models"
-
-export default interface Content {
-	feed: Feed
-	posts: Post[]
-}
