--- frontend/lib/feed/index.ts
@@ -1,6 +1,0 @@
-import type Content from "./Content"
-import fetchContent from "./fetchContent"
-import Feeds from "./Feeds"
-
-export default Feeds
-export { Content, fetchContent }
